+++ examples/pyplot/histo_3D.py
@@ -1,15 +1,17 @@
-"""Histogram in 3D.
-The size of each cube is proportional to the scalar value at that point"""
+"""Histogram (or plot) in 3D.
+The size of each cube is proportional to the value at that point"""
 import numpy as np
 from vedo import Volume, Cube, Glyph, show
 
+# Make up some arbitrary data
 X, Y, Z = np.mgrid[:4, :8, :8]
-scalar_field = 2-((X-4)**2 + (Y-4)**2 + (Z-4)**2)/25
+counts = 50 - ( (X-4)**2 + (Y-4)**2 + (Z-4)**2 )
 
-pcloud = Volume(scalar_field).topoints()
+# This is now a point cloud with an associated array of counts
+pcloud = Volume(counts).topoints()
 
-marker = Cube().scale(.4)
+marker = Cube().scale(0.015)
 glyphed_pcl = Glyph(pcloud, marker, scaleByScalar=True)
-glyphed_pcl.cmap('seismic').addScalarBar()
+glyphed_pcl.cmap('seismic').addScalarBar('counts')
 
 show(glyphed_pcl, __doc__, axes=1)